Add modified ghidra_plugin to autogenerate checks
When analyzing many binaries, I can in handy to just generate the results "on the fly".
Heavily based on the already existing plugin (import not possible as main would be executed), but doesn't bother the user.
Requires cwe_checker to be a callable executable on the shell of the user/host. (Note: the PATH env may not include your local rust binaries! I had to link that into my .local/bin. )